# Optimize serialization of []any and map[string]any