Improve serialization performance by specializing slices or maps of `any`, and optimizing reflective looping through maps.

Special-casing `[]any` and especially `map[string]any`, which are common parameter types, eliminates allocation overhead to loop through these values.

Using `MapRange` rather than `MapKeys` and `MapIndex` (and using `Value.SetIterKey/Value` rather than `MapIter.Key/Value`) significantly reduces allocations when looping through other map types.