@robsdedude 2025-01-31
FYI, I've been having a look and the speed-up looks very promising! I've not spotted any logic issues either :muscle: great job! But you've spiked my curiosity. I'm currently looking into whether there's a way to generalize this optimization so that it applies to even more cases.