@robsdedude 2025-01-31
I didn't succeed :/ everything I tried didn't improve the situation or made it even worse. What I've don, however, is added all `map[String]xyz` short paths that also exist for `[]xyz`.

Here are some benchmarks I quickly threw together to validate that this PR makes things faster:

```
### BEFORE ###

BenchmarkOutgoing
BenchmarkOutgoing/pack_[]any
BenchmarkOutgoing/pack_[]any-16                                      	 4242537	       250.7 ns/op
BenchmarkOutgoing/pack_[]bool
BenchmarkOutgoing/pack_[]bool-16                                     	 2663608	       455.2 ns/op
BenchmarkOutgoing/pack_map[string]any
BenchmarkOutgoing/pack_map[string]any-16                             	  943662	      1342 ns/op
BenchmarkOutgoing/pack_map[string]string
BenchmarkOutgoing/pack_map[string]string-16                          	 3308748	       375.1 ns/op
BenchmarkOutgoing/pack_map[string]any_(any_always_string)
BenchmarkOutgoing/pack_map[string]any_(any_always_string)-16         	  873051	      1281 ns/op
BenchmarkOutgoing/pack_map[string][]string
BenchmarkOutgoing/pack_map[string][]string-16                        	  803454	      1547 ns/op
BenchmarkOutgoing/pack_map[string]any_(any_always_[]string)
BenchmarkOutgoing/pack_map[string]any_(any_always_[]string)-16       	  825868	      1558 ns/op
BenchmarkOutgoing/pack_[]byte
BenchmarkOutgoing/pack_[]byte-16                                     	45481484	        34.66 ns/op


### AFTER ###

BenchmarkOutgoing
BenchmarkOutgoing/pack_[]any
BenchmarkOutgoing/pack_[]any-16                                      	 5936780	       178.7 ns/op
BenchmarkOutgoing/pack_[]bool
BenchmarkOutgoing/pack_[]bool-16                                     	 2564455	       472.9 ns/op
BenchmarkOutgoing/pack_map[string]any
BenchmarkOutgoing/pack_map[string]any-16                             	 3001713	       407.4 ns/op
BenchmarkOutgoing/pack_map[string]string
BenchmarkOutgoing/pack_map[string]string-16                          	 3258835	       386.6 ns/op
BenchmarkOutgoing/pack_map[string]any_(any_always_string)
BenchmarkOutgoing/pack_map[string]any_(any_always_string)-16         	 3530245	       374.7 ns/op
BenchmarkOutgoing/pack_map[string][]string
BenchmarkOutgoing/pack_map[string][]string-16                        	  857400	      1437 ns/op
BenchmarkOutgoing/pack_map[string]any_(any_always_[]string)
BenchmarkOutgoing/pack_map[string]any_(any_always_[]string)-16       	 1801963	       584.2 ns/op
BenchmarkOutgoing/pack_[]byte
BenchmarkOutgoing/pack_[]byte-16                                     	35928717	        39.06 ns/op
```