@zolstein 2025-02-01
Unfortunately, my experience is that it's _quite_ hard to optimize iterating a map with reflection - it's always much faster if you can statically convert to a known map type.

I did find a couple other optimizations, though I'm not sure if you'd want to take them.

The first is that I noticed that using MapIter.Value only allocates when the map values are not pointer-types. Therefore, you can branch on the map value type and avoid allocating a slot for the value if it has pointers, otherwise use the code as written.

```golang
	if isPointerType(t.Elem()) {
		r := v.MapRange()
		for r.Next() {
			key.SetIterKey(r)
			o.packer.String(key.String())
			o.packX(r.Value().Interface())
		}
	} else {
		value := reflect.New(t.Elem()).Elem()
		r := v.MapRange()
		for r.Next() {
			key.SetIterKey(r)
			value.SetIterValue(r)
			o.packer.String(key.String())
			o.packX(value.Interface())
		}
	}
```

The second is that you can put an extra string field in the `outgoing` struct, and use it as the place to store the key value. This lets you avoid an allocation per-map, which is especially useful if there are nested structures. However, the naive version of this breaks if the map has keys that are a new-type of string, and the only workaround I found for this required a bit of unsafe code.

```golang
	key := reflect.NewAt(t.Key(), unsafe.Pointer(&o.stringSlot)).Elem()
```