@zolstein 2025-02-01
Also, FYI, it's often helpful to include the `-benchmem` flag when benchmarking, so you also get data on allocations. That's especially true here, since allocations drive most of the performance difference.