@zolstein 2025-02-05
@robsdedude A note on your benchmark code - you're not resetting the buffer inside the Packer between benchmark iterations, which means the work done between iterations isn't consistent - some iterations will need to resize the buffer, others won't. Bytes allocated theoretically average out, though number of allocations won't. IMO it's better to pre-allocate a buffer big enough to avoid resizes and reset it between runs. That way, you can isolate the performance of the serialization code. You also might want to trigger a GC (and reset the timer) between benchmarks.

```golang
buffer := make([]byte, 0, 1024*1024)
...
	out := &outgoing{
		chunker:   newChunker(),
		packer:    packstream.Packer{},
		onPackErr: func(err error) { inner.Error(err) },
	}
	runtime.GC()
	inner.ResetTimer()
	for _i := 0; _i < inner.N; _i++ {
		out.packer.Begin(buffer)
		out.packX(load.data)
	}
```