@StephenCathcart 2025-02-13
Changes and tests look great, thanks for your help on this @zolstein 🙇